added skip pgp check for makepkg
This is useful when installing a new system and AUR packages need to be
installed before GnuPG is installed and configured (to automatically
retrieve public keys). E.g. cower requires PGP signature check and it needs
to be installed before pacaur can be installed and used for further AUR
package installation