Upgrade to latest actions versions
We need to use the latest versions because older nodejs runtimes have
been deprecated and removed some time ago.

As a bonus, the latest version of the setup-python action supports
caching pip dependencies out-of-the-box (setting the appropriate cache
option) so we can eliminate a few of the manual steps to setup
dependency caching.

Finally, I also bumped the python version to 3.13 which is what
currently ships in the Arch repositories, not that it really matters.